Rename `tools` to `tool_manager`
